mock renderer impl

+++ modules/graphics-hal/src/backends/empty/mod.rs
@@ -312,7 +312,7 @@ impl device::Device<Backend> for Device {
         Ok(())
     }
 
-    unsafe fn create_shader_module(&self, _: &[u32]) -> Result<(), device::ShaderError> {
+    unsafe fn create_shader_module(&self, _: &[u8]) -> Result<(), device::ShaderError> {
         Ok(())
     }
